+++ src/entities/product/ui/index.ts
@@ -1,3 +1,2 @@
 export { default as RankingItemCard } from './RankingItemCard';
 export { default as ProductOverview } from './ProductOverview';
-export { default as ProductTabs } from './ProductTabs';
