+++ src/widgets/index.ts
@@ -1,3 +1,4 @@
 export * from './navigation';
 export * from './themeSection';
-export * from './orderForm'; 
+export * from './orderForm';
+export * from './productTabs';
