+++ src/widgets/productTabs/components/index.ts
@@ -1,0 +1,3 @@
+export { default as DescriptionTabContent } from './DescriptionTabContent';
+export { default as ReviewTabContent } from './ReviewTabContent';
+export { default as DetailTabContent } from './DetailTabContent';
