+++ src/widgets/productTabs/styles.ts
